Modifier update.
Made application of modifiers possible now.

Health modifiers now work and can be applied to anything with a health script.

API functionality to remove the modifiers has been scripted but a way to call the API needs to be created.

Other modifiers to be announced.

A purely healthMultiplier attack has been created.