Started work on a wave system.
Started work on a wave system, does not work yet as it's late but I need to sleep.

Will work by letting you put what enemies you want in each wave to spawn and then it will spawn those enemies until there is no more which is when the wave will be over.