Got a basic wave spawner working.
Got a basic wave spawner working.

Whenever a wave is started each enemy is placed at a spawn point.

To start the next wave there is a function called StartNextWave() that can be put on a button I guess to start the next wave available. If no waves are available then it will set a boolean called "gameOver" to true.

That gameOver boolean can be accessed and checked with another public bool function called "IsGameOver()".